Now uses version + build, and more (see details)
1. XML_Feed now includes version and build
2. updated FILENAME=
3. release notes now won’t re-download if file already exists
4. added ‘localsha’
5. updated mount ‘nobrowse’

+++ di-encryptme.sh
@@ -1,11 +1,9 @@
 #!/usr/bin/env zsh -f
-# Purpose: Download and install the latest version of EncryptMe nee Cloak
+# Purpose: Download and install the latest version of EncryptMe (née Cloak)
 #
 # From:	Timothy J. Luoma
 # Mail:	luomat at gmail dot com
 # Date:	2015-11-06
-
-## 2018-07-10 - renamed di-cloak to di-encryptme to reflect new name
 
 NAME="$0:t:r"
 
@@ -26,22 +24,28 @@ fi
 
 XML_FEED='https://www.getcloak.com/updates/osx/public/'
 
-	# No other version info in feed
+	# 2020-06-05 - as of version 4.2.1 / 26 Feb 2020 the feed now shows both
+	# sparkle:shortVersionString and sparkle:version
 INFO=($(curl -sfL "$XML_FEED" \
 		| tr ' ' '\012' \
-		| egrep '^(url|sparkle:version)=' \
-		| tail -2 \
+		| egrep '^(url|sparkle:shortVersionString|sparkle:version)=' \
+		| tail -3 \
+		| sort \
 		| awk -F'"' '//{print $2}'))
 
-URL="$INFO[1]"
+	# 'sort' in the INFO= line should keep this consistent
+LATEST_VERSION="$INFO[1]"
 
-LATEST_VERSION="$INFO[2]"
+LATEST_BUILD="$INFO[2]"
 
-if [ "$INFO" = "" -o "$LATEST_VERSION" = "" -o "$URL" = "" ]
+URL="$INFO[3]"
+
+if [ "$INFO" = "" -o "$LATEST_VERSION" = ""  -o "$LATEST_BUILD" = "" -o "$URL" = "" ]
 then
 	echo "$NAME: Error: bad data received:
 	INFO: $INFO
 	LATEST_VERSION: $LATEST_VERSION
+	LATEST_BUILD: $LATEST_BUILD
 	URL: $URL
 	"
 
@@ -51,42 +55,64 @@ fi
 if [[ -e "$INSTALL_TO" ]]
 then
 
-	INSTALLED_VERSION=`defaults read "$INSTALL_TO/Contents/Info" CFBundleVersion`
+	INSTALLED_VERSION=$(defaults read "${INSTALL_TO}/Contents/Info" CFBundleShortVersionString)
 
-	if [[ "$LATEST_VERSION" == "$INSTALLED_VERSION" ]]
-	then
-		echo "$NAME: Up-To-Date ($INSTALLED_VERSION)"
-		exit 0
-	fi
+	INSTALLED_BUILD=$(defaults read "${INSTALL_TO}/Contents/Info" CFBundleVersion)
 
 	autoload is-at-least
 
 	is-at-least "$LATEST_VERSION" "$INSTALLED_VERSION"
 
-	if [ "$?" = "0" ]
+	VERSION_COMPARE="$?"
+
+	is-at-least "$LATEST_BUILD" "$INSTALLED_BUILD"
+
+	BUILD_COMPARE="$?"
+
+	if [ "$VERSION_COMPARE" = "0" -a "$BUILD_COMPARE" = "0" ]
 	then
-		echo "$NAME: Installed version ($INSTALLED_VERSION) is ahead of official version $LATEST_VERSION"
+		echo "$NAME: Up-To-Date ($INSTALLED_VERSION/$INSTALLED_BUILD)"
 		exit 0
 	fi
 
-	echo "$NAME: Outdated (Installed = $INSTALLED_VERSION vs Latest = $LATEST_VERSION)"
+	echo "$NAME: Outdated: $INSTALLED_VERSION/$INSTALLED_BUILD vs $LATEST_VERSION/$LATEST_BUILD"
 
+	FIRST_INSTALL='no'
+
+	if [[ ! -w "$INSTALL_TO" ]]
+	then
+		echo "$NAME: '$INSTALL_TO' exists, but you do not have 'write' access to it, therefore you cannot update it." >>/dev/stderr
+
+		exit 2
+	fi
+
+else
+
+	FIRST_INSTALL='yes'
 fi
 
-FILENAME="$HOME/Downloads/$INSTALL_TO:t:r-$LATEST_VERSION.dmg"
+FILENAME="$HOME/Downloads/${${INSTALL_TO:t:r}// /}-${${LATEST_VERSION}// /}_${${LATEST_BUILD}// /}.dmg"
 
 if (( $+commands[lynx] ))
 then
 
-	RELEASE_NOTES_URL="$XML_FEED"
+	if [[ -e "$FILENAME:r.txt" ]]
+	then
 
-	( echo -n "$NAME: Release Notes for ";
-		curl -sfL "$RELEASE_NOTES_URL" \
-		| sed 	-e "1,/<title>Encrypt.me $LATEST_VERSION<\/title>/d" \
-				-e '/<pubDate>/,$d' \
-				-e "s#<description><\!\[CDATA\[##g ; s#\]\]>##g" \
-		| lynx -dump -nomargins -width=10000 -assume_charset=UTF-8 -pseudo_inlines -stdin ;
-		echo "\nSource: XML_FEED <${RELEASE_NOTES_URL}>" ) | tee "$FILENAME:r.txt"
+		cat "$FILENAME:r.txt"
+
+	else
+
+		RELEASE_NOTES_URL="$XML_FEED"
+
+		( echo -n "$NAME: Release Notes for ";
+			curl -sfL "$RELEASE_NOTES_URL" \
+			| sed 	-e "1,/<title>Encrypt.me $LATEST_VERSION<\/title>/d" \
+					-e '/<pubDate>/,$d' \
+					-e "s#<description><\!\[CDATA\[##g ; s#\]\]>##g" \
+			| lynx -dump -nomargins -width=10000 -assume_charset=UTF-8 -pseudo_inlines -stdin ;
+			echo "\nSource: XML_FEED <${RELEASE_NOTES_URL}>" ) | tee "$FILENAME:r.txt"
+	fi
 
 fi
 
@@ -103,50 +129,74 @@ EXIT="$?"
 
 [[ ! -s "$FILENAME" ]] && echo "$NAME: $FILENAME is zero bytes." && rm -f "$FILENAME" && exit 0
 
-MNTPNT=$(hdiutil attach -nobrowse -plist "$FILENAME" 2>/dev/null \
-		| fgrep -A 1 '<key>mount-point</key>' \
-		| tail -1 \
-		| sed 's#</string>.*##g ; s#.*<string>##g')
-
-if [[ "$MNTPNT" == "" ]]
-then
-		echo "$NAME: Failed to mount $FILENAME. (MNTPNT is empty)"
-		exit 0
-fi
-
-echo "$MNTPNT"
-
-if [ -e "$INSTALL_TO" ]
-then
-		# Quit app, if running
-	pgrep -xq "EncryptMe" \
-	&& LAUNCH='yes' \
-	&& killall EncryptMe
-
-		# move installed version to trash
-	mv -vf "$INSTALL_TO" "$HOME/.Trash/EncryptMe.$INSTALLED_VERSION.app"
-fi
-
-echo "$NAME: Installing $FILENAME to $INSTALL_TO:h/"
-
-ditto --noqtn "$MNTPNT/EncryptMe.app" "$INSTALL_TO"
+egrep -q '^Local sha256:$' "$FILENAME:r.txt" 2>/dev/null
 
 EXIT="$?"
 
-if [ "$EXIT" = "0" ]
+if [ "$EXIT" = "1" -o ! -e "$FILENAME:r.txt" ]
 then
+	(cd "$FILENAME:h" ; \
+	echo "\n\nLocal sha256:" ; \
+	shasum -a 256 -p "$FILENAME:t" \
+	)  >>| "$FILENAME:r.txt"
+fi
 
-	echo "$NAME: Updated $INSTALL_TO"
 
+echo "$NAME: Mounting $FILENAME:"
+
+MNTPNT=$(hdiutil attach -nobrowse -plist "$FILENAME" 2>/dev/null \
+	| fgrep -A 1 '<key>mount-point</key>' \
+	| tail -1 \
+	| sed 's#</string>.*##g ; s#.*<string>##g')
+
+if [[ "$MNTPNT" == "" ]]
+then
+	echo "$NAME: MNTPNT is empty"
+	exit 1
 else
-	echo "$NAME: 'ditto' failed (\$EXIT = $EXIT)"
+	echo "$NAME: MNTPNT is $MNTPNT"
+fi
+
+if [[ -e "$INSTALL_TO" ]]
+then
+		# Quit app, if running
+	pgrep -xq "$INSTALL_TO:t:r" \
+	&& LAUNCH='yes' \
+	&& osascript -e "tell application \"$INSTALL_TO:t:r\" to quit"
+
+		# move installed version to trash
+	echo "$NAME: moving old installed version to '$TRASH'..."
+	mv -f "$INSTALL_TO" "$TRASH/$INSTALL_TO:t:r.${INSTALLED_VERSION}_${INSTALLED_BUILD}.app"
+
+	EXIT="$?"
+
+	if [[ "$EXIT" != "0" ]]
+	then
+
+		echo "$NAME: failed to move '$INSTALL_TO' to '$TRASH'. ('mv' \$EXIT = $EXIT)"
+
+		exit 1
+	fi
+fi
+
+echo "$NAME: Installing '$MNTPNT/$INSTALL_TO:t' to '$INSTALL_TO': "
+
+ditto --noqtn -v "$MNTPNT/$INSTALL_TO:t" "$INSTALL_TO"
+
+EXIT="$?"
+
+if [[ "$EXIT" == "0" ]]
+then
+	echo "$NAME: Successfully installed $INSTALL_TO"
+else
+	echo "$NAME: ditto failed"
 
 	exit 1
 fi
 
-diskutil eject "$MNTPNT"
+[[ "$LAUNCH" = "yes" ]] && open -a "$INSTALL_TO"
 
-[[ "$LAUNCH" == "yes" ]] && open -a "$INSTALL_TO"
+echo -n "$NAME: Unmounting $MNTPNT: " && diskutil eject "$MNTPNT"
 
 exit 0
 #EOF
